feat: convert to Typescript
